added NRPC to the EXPOSE Docker statements. It isn't needed when using publishing them with -p The EXPOSE statements are more for documentation purposes.
Optimized the management script. The info command does now only show the most important settings.
"inspect" or "info full" does show all information.
Querying information is time consuming because each information is a separate Docker command.
It slows down standard operations which don't need all the info.
So the routines to inspect the image and the container are now optimized.
This makes interaction with the script faster