a couple of bigger changes and additions first of all I renamed the "appdevpack" image to "protn" and added a "iam" image. neither the "proton" nor the "iam" dockerimage is complete. but they both are correctly build and copy the required data. the next step is to try to automate the installation.
in addition the configuration is now implemented in a way that all config files are first checked in /local/cfg and afterwards in the same directory of the binary.
this allows to keep configuration changes clean outside the git directory.

another lager change is that the docker containers all (but IAM currently) switch to the "notes" user before finishing installation.
that allows the container to start the entrypoing with "notes" instead of "root".
This security improvement needed a couple of changes in multiple scripts (start script, entry point script, ..).

Another change is the update to start script version 3.2.2, wich also comes with some new functonality and updates.
Some of the changes to switch from "root" to "notes" had to be implemented at start script level.